Se realiza consumo de api desplegada en heroku
