+++ app/src/main/java/com/uco/pdm/toolplus/integration/request/LoginRequest.kt
@@ -1,0 +1,3 @@
+package com.uco.pdm.toolplus.integration.request
+
+data class LoginRequest(val username: String, val password: String)
