+++ app/src/main/java/com/uco/pdm/toolplus/integration/response/TokenResponse.kt
@@ -1,0 +1,7 @@
+package com.uco.pdm.toolplus.integration.response
+
+import com.google.gson.annotations.SerializedName
+
+data class TokenResponse (
+    @SerializedName("token") var token: String
+)
